# api/api-docs.yaml
openapi: 3.0.0
info:
  title: Greenalytic User Management API
  version: 1.0.0
  description: API for managing users with authentication and pagination
servers:
  - url: http://localhost:5000/greenalytic
    description: Local Development Server
  - url: https://greenalytic-vehicle-monitoring-1.onrender.com/greenalytic
    description: Production Server

# Define Authentication Scheme
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  # JSON Web Token (JWT) format
...
paths:
...
  /users/{id}:
    delete:
      summary: Soft Delete a User
      tags: [users]
      security:
        - BearerAuth: []  # Requires authentication (JWT)
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the user to delete
          schema:
            type: string
      responses:
        '200':
          description: User deleted (soft delete)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "User deleted (soft delete)"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "User not found"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Something went wrong. Please try again."
    get:
      summary: Get User by ID
      tags: [users]
      security:
        - BearerAuth: []  # Requires authentication (JWT)
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the user to retrieve
          schema:
            type: string
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      username:
                        type: string
                      email:
                        type: string
                      phoneNumber:
                        type: string
                      role:
                        type: string
                      verified:
                        type: boolean
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "User not found"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Something went wrong. Please try again."
    patch:
      summary: Update User by ID
      tags: [users]
      security:
        - BearerAuth: []  # Requires authentication (JWT)
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the user to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "UpdatedUsername"
                email:
                  type: string
                  format: email
                  example: "updatedemail@example.com"
                phoneNumber:
                  type: string
                  example: "1234567890"
                role:
                  type: string
                  example: "admin"
                verified:
                  type: boolean
                  example: true
              additionalProperties: false  # Prevents sending unexpected fields
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "User updated successfully"
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      username:
                        type: string
                      email:
                        type: string
                      phoneNumber:
                        type: string
                      role:
                        type: string
                      verified:
                        type: boolean
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
        '400':
          description: Invalid request or update failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Failed to update user"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "User not found"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Something went wrong. Please try again."
...